requirements.api.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
torch==2.2.2+cpu
transformers>=4.30.0
python-dotenv==1.0.0
